Various changes:
- `free` should entries shouldn't be just one (list in out thread), they
  should be (list (list in out thread)) to handle multiple open
  connections.

- Convinced myself that we I need to guard `free` with a semaphore,
  after all.

- Convinced myself that we timeout's don't need to hash-remove! the item
  from `free`, instead can just close the ports.

- When pulling ports off `free`, need to check if the ports are still
  open. Need to do this anyway, and it has the side benefit of
  enabling the previous change.

Having said all that: As of this commit there still appears to be a
problem when running with current-pool-timeout set to 0. I don't
understand this because the code path in that case *should* be
equivalent to the old, non-pooling behavior. From the behavior, it's
obviously not identical, but I can't (yet) see how.

The problem is specific to www.yahoo.com. The error is

    custom-port-pipe-read: input port is closed