Remove microsoft.com from test sites; #12
It's not yet clear to me what the problem is. If it's simply Microsoft
blacklisting us as a "scraper", then this probably is sufficient to
close the issue. Otherwise, there might be some problem on our end we
should fix. For now, planning to merge this commit, regardless of
whether it turns out to be the last word, or, simply a temporary
mitigation.