Always supply a User-Agent request header; fixes #12
If user's request headers don't already supply a User-Agent, add one.

With this change, we can revert the change in the previous commit; in
other words we can add www.microsoft.com back to the list of test
sites.

Also, change most of the test sites to be https.

Also, refactor some of the test code so that it's easier to enter the
test submodule and try individual sites; that is, replace the
"test" (all sites) function with a "test-site" function.