Add details to check-no-acquire failure message
Include correct srcloc for test failures.

This is for issue #13. Although this does not fix the issue -- which I
can't seem to reproduce so far -- the next time it fails on DrDr this
should at least narrow down which of the tests is failing, which might
provide some clue what to do next.